+++ packages/xrpl/HISTORY.md
@@ -14,9 +14,10 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 ### Added
 - `getNFTokenID` lets you get the NFTokenID after minting an NFT
 
-
 ### Changed
 * All tests now use the Jest test runner and have been refactored for consistency across all packages
+* Updated faucet to for hooks to use v3 testnet
+* Removed nft-devnet faucet support as it has been decommissioned
 
 ### Deprecated
 Wallet.fromMmnemonic()